+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## dev
+- delete a single job
+- `all_jobs` shows all jobs in the `jobs/status` route.
 
 ## [0.0.2] - 2018-08-06
 - `__JOBID__` Placeholder